Polynomials4ML.OrthPolyBasis1D3T — Type

OrthPolyBasis1D3T: defines a basis of polynomials in terms of a 3-term recursion,

\[\begin{aligned} P_1(x) &= A_1 \\ P_2 &= A_2 x + B_2 \\ P_{n} &= (A_n x + B_n) P_{n-1}(x) + C_n P_{n-2}(x) \end{aligned}\]

Typically (but not necessarily) such bases are obtained by orthogonalizing the monomials with respect to a user-specified distribution, which can be either continuous or discrete but must have a density function. See also

legendre_basis

chebyshev_basis

jacobi_basis

Polynomials4ML.chebyshev_basis — Function

chebyshev_basis(N::Integer):

Constructs an OrthPolyBasis1D3T object representing a possibly rescaled version of the basis of Chebyshev polynomials of the first kind. N is the length of the basis, not the degree.

Careful: the normalisation may be non-standard.

Polynomials4ML.legendre_basis — Function

legendre_basis(N::Integer):

Constructs an OrthPolyBasis1D3T object representing a possibly rescaled version of the basis of Legendre polynomials (L2 orthonormal on [-1, 1]). N is the length of the basis, not the degree.

Careful: the normalisation may be non-standard.

Polynomials4ML.jacobi_basis — Function

jacobi_basis(N::Integer, α::Real, β::Real):

Constructs an OrthPolyBasis1D3T object representing a possibly rescaled version of the basis of Jacobi polynomials Jαβ. N is the length of the basis, not the degree.

Careful: the normalisation may be non-standard.

Polynomials4ML.MonoBasis — Type

Standard Monomials basis. This should very rarely be used. Possibly useful in combination with a transformation of the inputs, e.g. exponential.

Polynomials4ML.CTrigBasis — Type

Complex trigonometric polynomials up to degree N (inclusive). The basis is constructed in the order

[1, exp(im*θ), exp(-im*θ), exp(2im*θ), exp(-2im*θ), ..., 
                                exp(N*im*θ), exp(-N*im*θ) ]

where θ is input variable.

Polynomials4ML.RTrigBasis — Type

RTrigBasis(N):

Real trigonometric polynomials up to degree N (inclusive). The basis is ordered as

[1, cos(θ), sin(θ), cos(2θ), sin(2θ), ..., cos(Nθ), sin(Nθ) ]

where θ is input variable.

Polynomials4ML.CYlmBasis — Type

CYlmBasis(maxL, T=Float64):

Complex spherical harmonics; see tests to see how they are normalized, and idx2lm on how they are ordered. The ordering is not guarenteed to be semver-stable.

The input variable is normally an rr::SVector{3, T}. This rr need not be normalized (i.e. on the unit sphere). The derivatives account for this, i.e. they are valid even when norm(rr) != 1.

maxL : maximum degree of the spherical harmonics

T : type used to store the coefficients for the associated legendre functions

Polynomials4ML.RYlmBasis — Type

RYlmBasis(maxL, T=Float64):

Real spherical harmonics; see tests to see how they are normalized, and idx2lm on how they are ordered. The ordering is not guarenteed to be semver-stable.

The input variable is normally an rr::SVector{3, T}. This rr need not be normalized (i.e. on the unit sphere). The derivatives account for this, i.e. they are valid even when norm(rr) != 1.

maxL : maximum degree of the spherical harmonics

T : type used to store the coefficients for the associated legendre functions

Polynomials4ML.ALPolynomials — Type

ALPolynomials : an auxiliary datastructure for evaluating the associated Legendre functions used for the spherical and solid harmonics. Constructor:

ALPolynomials(maxL::Integer, T::Type=Float64)

This is not part of the public API and not guaranteed to be semver-stable. Only the resulting harmonics that use the ALPs are guaranteed to be backward compatible.

Important Note: evaluate_ed!` does NOT return derivatives, but rather produces rescaled derivatives for better numerical stability near the poles. See comments in code for details on how to use the ALP derivatives correctly.

Polynomials4ML.CRlmBasis — Type

complex spherical harmonics:

Yₗ⁰(θ, φ) = P̄ₗ⁰(cosθ)/√2 Yₗᵐ(θ, φ) = P̄ₗᵐ(cosθ)exp(imφ)/√2 Yₗ⁻ᵐ(θ, φ) = (-1)ᵐ P̄ₗᵐ(cosθ)/√2 exp(-imφ)

solid harmonics:

γₗᵐ(r, θ, φ) = rˡYₗᵐ(θ, φ)

Polynomials4ML.RRlmBasis — Type

real spherical harmonics:

Yₗ⁰ = P̄ₗ⁰/√2 Yₗᵐ = Re(P̄ₗᵐ(cosθ)/√2 exp(imφ)) Yₗ⁻ᵐ = -Im(P̄ₗᵐ(cosθ)/√2 exp(imφ))

solid harmonics:

Sₗ⁰ = √(4π/2l+1) rˡP̄ₗ⁰/√2 Sₗᵐ = (-1)ᵐ√(8π/2l+1) rˡ Re(P̄ₗᵐ(cosθ)/√2 exp(imφ)) Sₗ⁻ᵐ = (-1)ᵐ√(8π/2l+1) rˡIm(P̄ₗᵐ(cosθ)/√2 exp(imφ))

Polynomials4ML.SimpleProdBasis — Type

Naive implementation of the product basis, intended only for testing

Polynomials4ML.SparseSymmProdDAG — Method

Construct the DAG used to evaluate an AA basis and returns it as a SparseSymmProdDAG

Arguments

spec : AA basis specification, list of vectors of integers / indices pointing into A

Kwargs:

filter = _-> true :

verbose = false : print some information about the

Polynomials4ML.SphericalCoords — Type

struct SphericalCoords : a simple datatype storing spherical coordinates of a point (x,y,z) in the format (r, cosφ, sinφ, cosθ, sinθ). Use spher2cart and cart2spher to convert between cartesian and spherical coordinates.

Polynomials4ML._init_luxparams — Method

a fall-back method for initalparameters that all AbstractPoly4MLBasis should overload

Polynomials4ML.cRlm! — Method

evaluate complex spherical harmonics

Polynomials4ML.cRlm! — Method

evaluate complex spherical harmonics

Polynomials4ML.cRlm_ed! — Method

evaluate gradients of complex spherical harmonics

Polynomials4ML.cRlm_ed! — Method

evaluate gradients of complex spherical harmonics

Polynomials4ML.cYlm! — Method

evaluate complex spherical harmonics

Polynomials4ML.cYlm! — Method

evaluate complex spherical harmonics

Polynomials4ML.cYlm_ed! — Method

evaluate gradients of complex spherical harmonics

Polynomials4ML.cYlm_ed! — Method

evaluate gradients of complex spherical harmonics

Polynomials4ML.chebyshev_basis — Method

chebyshev_basis(N::Integer):

Constructs an OrthPolyBasis1D3T object representing a possibly rescaled version of the basis of Chebyshev polynomials of the first kind. N is the length of the basis, not the degree.

Careful: the normalisation may be non-standard.

Polynomials4ML.dspher_to_dcart — Method

convert a gradient with respect to spherical coordinates to a gradient with respect to cartesian coordinates

Polynomials4ML.idx2l — Method

Partial inverse of lm2idx: given an index into a vector of Ylm values, return the l index.

Polynomials4ML.idx2lm — Method

Inverse of lm2idx: given an index into a vector of Ylm values, return the l, m indices.

Polynomials4ML.index_p — Method

index_p(l,m): Return the index into a flat array of Associated Legendre Polynomials P_l^m for the given indices (l,m). P_l^m are stored in l-major order i.e.

[P(0,0), [P(1,0), P(1,1), P(2,0), ...]

Polynomials4ML.jacobi_basis — Method

jacobi_basis(N::Integer, α::Real, β::Real):

Constructs an OrthPolyBasis1D3T object representing a possibly rescaled version of the basis of Jacobi polynomials Jαβ. N is the length of the basis, not the degree.

Careful: the normalisation may be non-standard.

Polynomials4ML.legendre_basis — Method

legendre_basis(N::Integer):

Constructs an OrthPolyBasis1D3T object representing a possibly rescaled version of the basis of Legendre polynomials (L2 orthonormal on [-1, 1]). N is the length of the basis, not the degree.

Careful: the normalisation may be non-standard.

Polynomials4ML.lm2idx — Method

lm2idx(l,m): Return the index into a flat array of real spherical harmonics Y_lm for the given indices (l,m). Y_lm are stored in l-major order i.e.

[Y(0,0), Y(1,-1), Y(1,0), Y(1,1), Y(2,-2), ...]

Polynomials4ML.lux — Method

lux(basis) : convert a basis / embedding object into a lux layer. This assumes that the basis accepts a number or short vector as input and produces an output that is a vector. It also assumes that batched operations are implemented, as well as some other functionality.

Polynomials4ML.maxL — Method

max L degree for which the alp coefficients have been precomputed

Polynomials4ML.rRlm! — Method

evaluate real solid harmonics

Polynomials4ML.rRlm_ed! — Method

evaluate gradients of real spherical harmonics

Polynomials4ML.rYlm! — Method

evaluate real spherical harmonics

Polynomials4ML.rYlm_ed! — Method

evaluate gradients of real spherical harmonics

Polynomials4ML.sizeP — Method

sizeP(maxL): Return the size of the set of Associated Legendre Polynomials $P_l^m(x)$ of degree less than or equal to the given maximum degree

Polynomials4ML.sizeY — Method

sizeY(maxL): Return the size of the set of spherical harmonics $Y_{l,m}(θ,φ)$ of degree less than or equal to the given maximum degree maxL